pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django-bootstrap5>=24.3",
    "django>=5.1.4",
    "gunicorn>=23.0.0",
    "uvicorn>=0.34.0",
    "whitenoise>=6.8.2",
    "uv>=0.5.12",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "django-filter>=24.3",
]
